+++ src/hm/hmRadio.h
@@ -16,6 +16,11 @@
 
 const char* const rf24AmpPowerNames[] = {"MIN", "LOW", "HIGH", "MAX"};
 
+#define TX_REQ_DREDCONTROL  0x50
+#define DRED_A5 0xa5
+#define DRED_5A 0x5a
+#define DRED_AA 0xaa
+#define DRED_55 0x55
 
 //-----------------------------------------------------------------------------
 // HM Radio class
@@ -57,6 +62,7 @@ class HmRadio : public Radio {
             pinMode(irq, INPUT_PULLUP);
 
             generateDtuSn();
+            DTU_RADIO_ID = ((uint64_t)(((mDtuSn >> 24) & 0xFF) | ((mDtuSn >> 8) & 0xFF00) | ((mDtuSn << 8) & 0xFF0000) | ((mDtuSn << 24) & 0xFF000000)) << 8) | 0x01;
 
             #ifdef ESP32
                 #if CONFIG_IDF_TARGET_ESP32C3 || CONFIG_IDF_TARGET_ESP32S2 || CONFIG_IDF_TARGET_ESP32S3
@@ -81,7 +87,7 @@ class HmRadio : public Radio {
             mNrf24.enableDynamicPayloads();
             mNrf24.setCRCLength(RF24_CRC_16);
             mNrf24.setAddressWidth(5);
-            mNrf24.openReadingPipe(1, reinterpret_cast<uint8_t*>(&mDtuSn));
+            mNrf24.openReadingPipe(1, reinterpret_cast<uint8_t*>(&DTU_RADIO_ID));
 
             // enable all receiving interrupts
             mNrf24.maskIRQ(false, false, false);
@@ -139,7 +145,7 @@ class HmRadio : public Radio {
             return mNrf24.isChipConnected();
         }
 
-        void sendControlPacket(Inverter<> *iv, uint8_t cmd, uint16_t *data, bool isRetransmit, bool isNoMI = true, bool is4chMI = false) {
+        void sendControlPacket(Inverter<> *iv, uint8_t cmd, uint16_t *data, bool isRetransmit, bool isNoMI = true, uint16_t powerMax = 0) {
             DPRINT(DBG_INFO, F("sendControlPacket cmd: 0x"));
             DBGHEXLN(cmd);
             initPacket(iv->radioId.u64, TX_REQ_DEVCONTROL, SINGLE_FRAME);
@@ -155,31 +161,64 @@ class HmRadio : public Radio {
                 }
             } else { //MI 2nd gen. specific
                 switch (cmd) {
+                    case Restart:
                     case TurnOn:
-                        //mTxBuf[0] = 0x50;
-                        mTxBuf[9] = 0x55;
-                        mTxBuf[10] = 0xaa;
+                        mTxBuf[9]  = DRED_55;
+                        mTxBuf[10] = DRED_AA;
                         break;
                     case TurnOff:
-                        mTxBuf[9] = 0xaa;
-                        mTxBuf[10] = 0x55;
+                        mTxBuf[9]  = DRED_AA;
+                        mTxBuf[10] = DRED_55;
                         break;
                     case ActivePowerContr:
-                        mTxBuf[9] = 0x5a;
-                        mTxBuf[10] = 0x5a;
-                        //Testing only! Original NRF24_DTUMIesp.ino code #L612-L613:
-                        //UsrData[0]=0x5A;UsrData[1]=0x5A;UsrData[2]=100;//0x0a;// 10% limit
-                        //UsrData[3]=((Limit*10) >> 8) & 0xFF;   UsrData[4]= (Limit*10)  & 0xFF;   //WR needs 1 dec= zB 100.1 W
-                        if (is4chMI) {
-                            mTxBuf[cnt++] = 100; //10% limit, seems to be necessary to send sth. at all, but for MI-1500 this has no effect
-                            //works (if ever!) only for absulute power limits!
-                            mTxBuf[cnt++] = ((data[0] * 10) >> 8) & 0xff; // power limit
-                            mTxBuf[cnt++] = ((data[0] * 10)     ) & 0xff; // power limit
-                        } else {
-                            mTxBuf[cnt++] = data[0]*10; // power limit
+                        if (data[1]<256) { // non persistent
+                            mTxBuf[9]  = DRED_5A;
+                            mTxBuf[10] = DRED_5A;
+                            //Testing only! Original NRF24_DTUMIesp.ino code #L612-L613:
+                            //UsrData[0]=0x5A;UsrData[1]=0x5A;UsrData[2]=100;//0x0a;// 10% limit
+                            //UsrData[3]=((Limit*10) >> 8) & 0xFF;   UsrData[4]= (Limit*10)  & 0xFF;   //WR needs 1 dec= zB 100.1 W
+                            if (!data[1]) {   //     AbsolutNonPersistent
+                                mTxBuf[++cnt] = 100; //10% limit, seems to be necessary to send sth. at all, but for MI-1500 this has no effect
+                                //works (if ever!) only for absulute power limits!
+                                mTxBuf[++cnt] = ((data[0] * 10) >> 8) & 0xff; // power limit in W
+                                mTxBuf[++cnt] = ((data[0] * 10)     ) & 0xff; // power limit in W
+                            } else if (powerMax) {       //relative, but 4ch-MI (if ever) only accepts absolute values
+                                mTxBuf[++cnt] = data[0]; // simple power limit in %, might be necessary to multiply by 10?
+                                mTxBuf[++cnt] = ((data[0] * 10 * powerMax) >> 8) & 0xff; // power limit
+                                mTxBuf[++cnt] = ((data[0] * 10 * powerMax)     ) & 0xff; // power limit
+                            } else {   // might work for 1/2ch MI (if ever)
+                                mTxBuf[++cnt] = data[0]; // simple power limit in %, might be necessary to multiply by 10?
+                            }
+                        } else {       // persistent power limit needs to be translated in DRED command (?)
+                            /* DRED instruction
+                            Order	 Function
+                            0x55AA	 Boot without DRM restrictions
+                            0xA5A5	 DRM0 shutdown
+                            0x5A5A	 DRM5 power limit 0%
+                            0xAA55	 DRM6 power limit 50%
+                            0x5A55	 DRM8 unlimited power operation
+                            */
+                            mTxBuf[0] = TX_REQ_DREDCONTROL;
+
+                            if (data[1] == 256UL) {   //     AbsolutPersistent
+                                if (data[0] == 0 && !powerMax) {
+                                    mTxBuf[9]  = DRED_A5;
+                                    mTxBuf[10] = DRED_A5;
+                                } else if (data[0] == 0 || !powerMax || data[0] < powerMax/4 ) {
+                                    mTxBuf[9]  = DRED_5A;
+                                    mTxBuf[10] = DRED_5A;
+                                } else if (data[0] <=  powerMax/4*3) {
+                                    mTxBuf[9]  = DRED_AA;
+                                    mTxBuf[10] = DRED_55;
+                                } else if (data[0] <=  powerMax) {
+                                    mTxBuf[9]  = DRED_5A;
+                                    mTxBuf[10] = DRED_55;
+                                } else if (data[0] > powerMax*2) {
+                                    mTxBuf[9]  = DRED_55;
+                                    mTxBuf[10] = DRED_AA;
+                                }
+                            }
                         }
-
-
                         break;
                     default:
                         return;
@@ -202,7 +241,7 @@ class HmRadio : public Radio {
         std::queue<packet_t> mBufCtrl;
 
     private:
-            bool getReceived(void) {
+        inline bool getReceived(void) {
             bool tx_ok, tx_fail, rx_ready;
             mNrf24.whatHappened(tx_ok, tx_fail, rx_ready); // resets the IRQ pin to HIGH
 
@@ -219,10 +258,10 @@ class HmRadio : public Radio {
                     if (p.packet[0] != 0x00) {
                         mBufCtrl.push(p);
                         if (p.packet[0] == (TX_REQ_INFO + ALL_FRAMES))  // response from get information command
-                                isLastPackage = (p.packet[9] > ALL_FRAMES); // > ALL_FRAMES indicates last packet received
+                            isLastPackage = (p.packet[9] > ALL_FRAMES); // > ALL_FRAMES indicates last packet received
                         else if (p.packet[0] == ( 0x0f + ALL_FRAMES) )  // response from MI get information command
-                                isLastPackage = (p.packet[9] > 0x10);       // > 0x10 indicates last packet received
-                            else if ((p.packet[0] != 0x88) && (p.packet[0] != 0x92)) // ignore fragment number zero and MI status messages //#0 was p.packet[0] != 0x00 &&
+                            isLastPackage = (p.packet[9] > 0x10);       // > 0x10 indicates last packet received
+                        else if ((p.packet[0] != 0x88) && (p.packet[0] != 0x92)) // ignore fragment number zero and MI status messages //#0 was p.packet[0] != 0x00 &&
                             isLastPackage = true;                       // response from dev control command
                     }
                 }
@@ -232,7 +271,7 @@ class HmRadio : public Radio {
         }
 
         void sendPacket(Inverter<> *iv, uint8_t len, bool isRetransmit, bool appendCrc16=true) {
-            updateCrcs(len, appendCrc16);
+            updateCrcs(&len, appendCrc16);
 
             // set TX and RX channels
             mTxChIdx = (mTxChIdx + 1) % RF_CHANNELS;
@@ -262,6 +301,7 @@ class HmRadio : public Radio {
             return iv->radioId.u64;
         }
 
+        uint64_t DTU_RADIO_ID;
         uint8_t mRfChLst[RF_CHANNELS];
         uint8_t mTxChIdx;
         uint8_t mRxChIdx;